reorg, refresh and upate
