+++ .github/ISSUE_TEMPLATE
@@ -1,0 +1,10 @@
+
+These samples and workflows are not officially supported. We will attempt to answer questions, but cannot gaurantee a timely response. For issues with the core software (ArcGIS Enterprise, ArcGIS Online, etc), please log a [support request](https://support.esri.com/en/contact-tech-support) or post on the [Esri Community](https://community.esri.com/). 
+
+### Question, Issue, or Enhancement?
+Do you have a question on a sample or workflow?
+Is something not working?
+Are you suggesting a new sample or workflow not listed?
+
+### Description
+Please provide as many details as possible
