+++ 3rdParty/IFTTT/README.md
@@ -1,0 +1,15 @@
+# [IFTTT](https://ifttt.com)
+IF This Then That provides a webhook connector. You can use the connector to listen for webhook requests from ArcGIS Portal and ArcGIS Online.
+
+### Quick Start
+
+Create a new Applet.
+
+1. Select **If This Add**
+2. Search "`webhook`" and select **Webhooks**
+3. Choose **Receive a web request with a JSON payload**
+4. Provide a name and **Create** the trigger
+5. Finish creating your applet with **Then That**
+6. **Continue** and **Finish**
+7. Find and retrieve the URL from [https://ifttt.com/maker_webhooks/settings](https://ifttt.com/maker_webhooks/settings)
+8. Copy and use the URL to setup the webhook within ArcGIS Portal or ArcGIS Online Hosted Feature Services
