+++ 3rdParty/README.md
@@ -1,0 +1,17 @@
+# 3rd Party Receivers
+Some of the popular vendors offering webhook services. Each link below provides a quick start guide to creating a webhook. The workflow must be finished by adding a trigger which responds to the webhook event by performing an action. As each vendor has various supported actions and different paths and business rules for these actions, they are not documented here. Please consult the respective documentation for each vendor to finish your workflow.
+
+
+### [IFTTT](/IFTTT)
+
+
+### [Make](/Make)
+
+
+### [Power Automate](/PowerAutomate)
+
+
+### [Tray.IO](/Tray.IO)
+
+
+### [Zapier](/Zapier)
