+++ 3rdParty/Tray.IO/README.md
@@ -1,0 +1,11 @@
+# [Tray.IO](https://www.tray.io)
+
+Tray.IO can listen for a webhook payload and perform an action.
+
+### Quick Start
+Tray.IO has [comprehensive documenation](https://tray.io/documentation/connectors/service/arcgis/) for ArcGIS Online Feature Services. Their ready to go ArcGIS connector will automatically create a webhook on the given Hosted Feature Service. After setting up the connection, it's simply a matter of completing your workflow. The [example provided](https://tray.io/documentation/connectors/service/arcgis/#example-usage) by Tray.IO demonstrates how the service could apply an edit to a hosted service.
+
+Using Tray.IO as a provider for ArcGIS Enterprise (Portal webhooks)
+write up how to create a webhook
+
+https://tray.io/documentation/connectors/triggers/webhook-trigger/
