+++ 3rdParty/Zapier/README.md
@@ -1,0 +1,14 @@
+# [Zapier](https://zapier.com)
+Zapier provides a webhook connector to build zaps (workflows).
+
+### Quick Start
+
+Create a new zap.
+
+1. Click **Create Zap** 
+2. Search for  `webhook` in the  **App Event** box
+3. Select **Webhooks by Zapier**
+4. For the **Trigger Event** select **Catch Hook**.
+5. Choose **Continue**
+6. Copy the URL for the **Custom Webhook URL** and configure your webhook within ArcGIS Portal or ArcGIS Online Hosted Feature Service.
+7. **Continue**. Click the **+ button** to add more options, such as **Format** or move to **Actions** to finish creating your zap.
