+++ Developer/javascript/node.js/server.js
@@ -1,4 +1,4 @@
-/*Copyright 2021 Esri
+/*Copyright 2019 Esri
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
    You may obtain a copy of the License at
