--- sample-workflow/config.py
@@ -1,13 +1,0 @@
-# Slack Token
-slackToken = "Bot User Oauth Access token"
-
-# Fully qualified domain name of Portal
-portalFQDN = "Portal URL"
-portalUsername = "Admin Username"
-portalPassword = "Admin Password"
-
-# item ID that portal should look for. 
-#searchItemID = "itemID"
-
-## Required Tags
-standardTags = ['Your', 'Organizations', 'standard', 'tags']
