+++ sample-workflows/README.md
@@ -1,0 +1,11 @@
+# Sample Workflows
+The samples listed here attempt to offer an almost complete workflow. With a few updates to make the sample specific to your environment, they provide solutions to some popular webhook - automated specific workflows. These sample workflows provide one way to achieve a goal and generally the same goal can be achieved in many different ways using different technologies or methods. Please follow the readme in the respective folders for deployment information.
+
+
+### [Write payloads from an Azure Function to an Azure Table storage](/azure-write-table)
+
+
+### [Send payloads via email from a Python Flask server](/python-email)
+
+
+### [Post item metadata to a Slack channel after parsing with the Python API](/slack)
