+++ sample-workflows/python-email/webhookListener.py
@@ -1,0 +1,45 @@
+"""  Copyright 2019 Esri
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License. """
+
+from flask import Flask, request
+from OpenSSL import SSL
+import json
+import sendEmail
+
+# File that webhook payloads will be written
+filename = 'C:\\temp\\webhookPayloads.txt' 
+
+app = Flask(__name__)
+
+@app.route('/', methods=['POST','GET'])
+def index():
+	if request.method == 'GET':
+		return '<h1>Hello from Webhook Listener!</h1>'
+
+	if request.method == 'POST':		
+		# Get the payload
+		req_data = request.get_json()
+		payload = json.dumps(req_data)
+		
+		# Write payload to file
+		with open(filename, 'a+') as f:    		
+			f.write(payload + '\n')
+
+		# Send email notification
+		sendEmail.send(payload)
+
+		return '{"success":"true"}'
+
+if __name__ == "__main__":   
+	# Certificate and key file. Cannot be self signed certs
+	context = ('ssl.cert', 'ssl.key')
+	# Listen on port 5000
+	app.run(host='0.0.0.0', port=5000, ssl_context=context, threaded=True, debug=True)
